// lib/api/schemas/create-user-dto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const createUserDtoRoleEnum = {
	ADMIN: "ADMIN",
	USER: "USER",
} as const;

export type CreateUserDtoRoleEnumKey =
	(typeof createUserDtoRoleEnum)[keyof typeof createUserDtoRoleEnum];

export type CreateUserDto = {
	/**
	 * @type string
	 */
	name: string;
	/**
	 * @type string
	 */
	email: string;
	/**
	 * @type string
	 */
	password: string;
	/**
	 * @type string
	 */
	role: CreateUserDtoRoleEnumKey;
};